src/test/java/com/leetcode/strings/easy/PathCrossing_1496Test.java
package com.leetcode.strings.easy;

import org.junit.jupiter.api.extension.ExtensionContext;
import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.Arguments;
import org.junit.jupiter.params.provider.ArgumentsProvider;
import org.junit.jupiter.params.provider.ArgumentsSource;

import java.util.stream.Stream;

import static org.assertj.core.api.Assertions.assertThat;
import static org.junit.jupiter.params.provider.Arguments.arguments;

class PathCrossing_1496Test {
  private final PathCrossing_1496 solution = new PathCrossing_1496();

  private static class TestArgumentsProvider implements ArgumentsProvider {
    @Override
    public Stream<? extends Arguments> provideArguments(ExtensionContext context) {
      return Stream.of(arguments("NES", false), arguments("NESWW", true));
    }
  }

  @ParameterizedTest
  @ArgumentsSource(TestArgumentsProvider.class)
  void isPathCrossing(String str, boolean expectedResult) {
    boolean actualResult = solution.isPathCrossing(str);

    assertThat(actualResult).isEqualTo(expectedResult);
  }
}
